+++ projects/minecraft/v1.19.2/vanilla/src/main/java/kittify/vanilla/common/integration/trinkets/KittifyTrinketsIntegration.java
@@ -9,5 +9,6 @@ public class KittifyTrinketsIntegration {
 
     public static void afterRegisterItems() {
         TrinketsApi.registerTrinket(KittifyItems.CUTE_COLLAR, new CuteCollarTrinket());
+        TrinketsApi.registerTrinket(KittifyItems.EXE_ARC, new EXEArcTrinket());
     }
 }