+++ CHANGELOG.md
@@ -9,6 +9,9 @@
 - Added Cute Collar (trinket/curio). (1.19.2)
     - 2x Movement Speed.
     - Does not drop on death.
+- Added EXE-Arc (trinket/curio). (1.19.2)
+    - Heals to full in 20 seconds.
+    - Does not drop on death.
 
 ## [0.0.6] - 2023-05-02